Added Seen Words
Well not really seen but able to be seen. The new table is to keep track
of which words the user is being exposed to. When logged in, the quizzes
now only choose from words the user has seen. The initial words are
added on account creation.

Eventually upon successfully answering the quiz for a word for the first
time, new words would be added to seen. This allows for a spaced
repetition of sorts. Right now, it simply cycles through the initial set
I assign and adds no new words.