Added Results and Expose New Words On Success
When logged in, the user only will be quized on words they have "seen".
When they get a newly seen word correct for the first time, they will
have a new word added to their seen list.

New words will be added until the full list of words is exposed.