Added Word Information and Basic User Statistics
The hsk word pages now include a table of the words in that level. If a
user is logged in, only the "seen" words appear in the list. The user's
list also includes the correct / incorrect guesses count for each seen
word.

A link was added between the word hsk page and the quiz pages.