Added Simplified or Traditional Option
I added a key to the session which keeps track of simplified or
traditional mode. The key's value determines what word is used for the
quiz and the user answer (simplified or traditional characters). It also
has some cosmetic effects.

Users' seen and result tables were modified to keep track of simplified
and traditional progression separately.

Fixed a bug involving Quiz Everything and viewing the words list.