Made Results Table Dependent on Seen, Added Timestamps
In adding different quiz types for simplified and traditional, and the
planned word or meaning option, I decided to have the results table look
at a seen word for user, word, quiz type. This changed a number of
queries.

I added timestamps to user creation, word appearance, and quiz question
and answer times. They're not utilized in the UI.

Also, I removed the ability and code around quizzing all words as a
logged in user since it messes up adding newly seen words in hsk levels.