Added Weighted Choice for Quiz Questions
A new `weight` column was added to `seen` which determines the frequency
of appearance of various words based on recent correct or incorrect
answers. Each word starts at weight 100, will increase or decrease by 10
on each answer, and has a minimum value of 50 and max value of 150.

The idea is to have some spaced repetition based on a user's
familiarity, but not bombard them with a single word or ignore a
well-known word.