Removed Navbar Collapse, JS, and Modified Pill Nav Style
Since the navbar was uncluttered, I removed the collapse functionality.
I believe that was the only thing using js, so I removed the bootstrap
js include.

It seems I missed that the pill nav on vertical mobile was squished. I
modified it to fit in smaller areas, so hopefully this deals with it.