Added Meta Tags and Footer Placeholder
I modified base.html to allow for a custom description meta tag, as well
as meta tags defined by a child (specifically nofollow on quiz/answer).
I tried to keep the non-user page titles and descriptions unique so the
search engines are happy.

I added a footer which I intend to fill in with help and contact info
type stuff.