Added Help and About Pages and Added Welcome Text
I added help and about pages to the footer. About contains information
about why I made Midki. Help acts as an FAQ to help guide the user in
what's going on.

I also added some welcome text to the main page when the user is not
logged in. It directs the user to the most basic functionlity and to
look at help or create an account.